+++ frontend/src/App.tsx
@@ -1,25 +1,45 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css"; // Importante para estilos del toast
 import "./App.css";
 
 // Importación de las páginas que se van a renderizar según la ruta
 import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
-import Navbar from "./components/NavBar"; 
+import DashboardPage from "./pages/DashboardPage"; // <-- tu futura vista protegida
+
+// Importación componentes
+import NavBar from "./components/NavBar"; 
+import PrivateRoute from "./components/PrivateRoute"; // <-- nuevo componente
+
+
+// Contexto global
+import { AuthProvider } from "./context/AuthProvider";
+
 
 function App() {
   return (
-    // Router principal que envuelve toda la aplicación
-    <Router>
-      <Navbar />
-      {/* Contenedor de las rutas */}
-      <Routes>
-        <Route path="/" element={<LoginPage />} />
-        <Route path="/register" element={<RegisterPage />} />
-      </Routes>
-      {/*Contenedor global de notificaciones */}
-      <ToastContainer position="top-right" autoClose={3000} />
-    </Router>
+    <AuthProvider>
+      <Router>
+        <NavBar />
+        <Routes>
+          {/* Públicas */}
+          <Route path="/" element={<LoginPage />} />
+          <Route path="/register" element={<RegisterPage />} />
+
+          {/* Protegidas */}
+          <Route
+            path="/dashboard"
+            element={
+              <PrivateRoute>
+                <DashboardPage />
+              </PrivateRoute>
+            }
+          />
+        </Routes>
+        <ToastContainer position="top-right" autoClose={3000} />
+      </Router>
+    </AuthProvider>
   );
 }
