+++ frontend/src/main.tsx
@@ -4,14 +4,15 @@ import "bootstrap/dist/css/bootstrap.min.css"; // CDN Bootstrap
 import "bootstrap/dist/js/bootstrap.bundle.min.js"; // Esto activa el colapso del menú
 import "react-toastify/dist/ReactToastify.css"; // Notificaciones
 import "bootstrap-icons/font/bootstrap-icons.css"; // Iconos de Bootstrap
+
 import "./App.css";
+
 import App from "./App.tsx";
 
 // Crea el punto de entrada de la aplicación React en el elemento con id="root"
 createRoot(document.getElementById("root")!).render(
-  // StrictMode: herramienta de desarrollo de React para detectar problemas potenciales
+  // StrictMode: Detecta problemas, se mantiene en entorno de desarrollo.
   <StrictMode>
-    {/* Se renderiza el componente principal de la aplicación */}
-    <App />
+      <App /> {/* Se renderiza el componente principal de la aplicación */}
   </StrictMode>
 );
